+++ cmd/copy.go
@@ -4,7 +4,8 @@ import (
 	"github.com/fiwon123/crower/internal/app"
 	"github.com/fiwon123/crower/internal/domain/copydom"
 	"github.com/fiwon123/crower/internal/errors"
-	cmdsHelper "github.com/fiwon123/crower/internal/helper/cmds"
+	"github.com/fiwon123/crower/internal/helper"
+
 	"github.com/spf13/cobra"
 )
 
@@ -22,7 +23,7 @@ Examples:
 `,
 	Run: func(cmd *cobra.Command, args []string) {
 
-		cfgFilePath, _ := cmdsHelper.GetPersistentConfigFlag(cmd)
+		cfgFilePath, _ := helper.GetPersistentConfigFlag(cmd)
 
 		app := app.InitApp(cfgFilePath)
