+++ internal/domain/opendom/hadler_test.go
@@ -6,13 +6,12 @@ import (
 
 	"github.com/fiwon123/crower/internal/domain/createdom"
 	"github.com/fiwon123/crower/internal/domain/opendom"
-
-	testshelper "github.com/fiwon123/crower/internal/helper/tests"
+	"github.com/fiwon123/crower/internal/helper"
 )
 
 func TestOpenFile(t *testing.T) {
 	t.Run("Open File single name", func(t *testing.T) {
-		app, _, err := testshelper.InitCrowerTests()
+		app, _, err := helper.InitCrowerTests()
 		if err != nil {
 			t.Fatalf("error before test: %v", err)
 		}
@@ -34,7 +33,7 @@ func TestOpenFile(t *testing.T) {
 	})
 
 	t.Run("Open File name with space", func(t *testing.T) {
-		app, _, err := testshelper.InitCrowerTests()
+		app, _, err := helper.InitCrowerTests()
 		if err != nil {
 			t.Fatalf("error before test: %v", err)
 		}
@@ -58,7 +57,7 @@ func TestOpenFile(t *testing.T) {
 
 func TestOpenFolder(t *testing.T) {
 	t.Run("Open c single name", func(t *testing.T) {
-		app, _, err := testshelper.InitCrowerTests()
+		app, _, err := helper.InitCrowerTests()
 		if err != nil {
 			t.Fatalf("error before test: %v", err)
 		}
@@ -80,7 +79,7 @@ func TestOpenFolder(t *testing.T) {
 	})
 
 	t.Run("Open Folder name with space", func(t *testing.T) {
-		app, _, err := testshelper.InitCrowerTests()
+		app, _, err := helper.InitCrowerTests()
 		if err != nil {
 			t.Fatalf("error before test: %v", err)
 		}
@@ -103,7 +102,7 @@ func TestOpenFolder(t *testing.T) {
 }
 
 func TestOpenSystem(t *testing.T) {
-	app, _, err := testshelper.InitCrowerTests()
+	app, _, err := helper.InitCrowerTests()
 	if err != nil {
 		t.Fatalf("error before test: %v", err)
 	}
