+++ internal/domain/searchdom/handler_test.go
@@ -5,12 +5,12 @@ import (
 	"testing"
 
 	"github.com/fiwon123/crower/internal/domain/searchdom"
-	testshelper "github.com/fiwon123/crower/internal/helper/tests"
+	"github.com/fiwon123/crower/internal/helper"
 )
 
 func TestSearchBrowser(t *testing.T) {
 	t.Run("Search Browser single name", func(t *testing.T) {
-		app, _, err := testshelper.InitCrowerTests()
+		app, _, err := helper.InitCrowerTests()
 		if err != nil {
 			t.Fatalf("error before test: %v", err)
 		}
@@ -24,7 +24,7 @@ func TestSearchBrowser(t *testing.T) {
 	})
 
 	t.Run("Search Browser name with space", func(t *testing.T) {
-		app, _, err := testshelper.InitCrowerTests()
+		app, _, err := helper.InitCrowerTests()
 		if err != nil {
 			t.Fatalf("error before test: %v", err)
 		}
@@ -40,7 +40,7 @@ func TestSearchBrowser(t *testing.T) {
 
 func TestSearchFile(t *testing.T) {
 	t.Run("Search File single name", func(t *testing.T) {
-		app, _, err := testshelper.InitCrowerTests()
+		app, _, err := helper.InitCrowerTests()
 		if err != nil {
 			t.Fatalf("error before test: %v", err)
 		}
@@ -54,7 +54,7 @@ func TestSearchFile(t *testing.T) {
 	})
 
 	t.Run("Search File name with space", func(t *testing.T) {
-		app, _, err := testshelper.InitCrowerTests()
+		app, _, err := helper.InitCrowerTests()
 		if err != nil {
 			t.Fatalf("error before test: %v", err)
 		}
@@ -70,7 +70,7 @@ func TestSearchFile(t *testing.T) {
 
 func TestSearchFolder(t *testing.T) {
 	t.Run("Search Folder single name", func(t *testing.T) {
-		app, _, err := testshelper.InitCrowerTests()
+		app, _, err := helper.InitCrowerTests()
 		if err != nil {
 			t.Fatalf("error before test: %v", err)
 		}
@@ -84,7 +84,7 @@ func TestSearchFolder(t *testing.T) {
 	})
 
 	t.Run("Search Folder name with space", func(t *testing.T) {
-		app, _, err := testshelper.InitCrowerTests()
+		app, _, err := helper.InitCrowerTests()
 		if err != nil {
 			t.Fatalf("error before test: %v", err)
 		}
@@ -100,7 +100,7 @@ func TestSearchFolder(t *testing.T) {
 
 func TestSearchFileAndFolder(t *testing.T) {
 	t.Run("Search Folder and File single name", func(t *testing.T) {
-		app, _, err := testshelper.InitCrowerTests()
+		app, _, err := helper.InitCrowerTests()
 		if err != nil {
 			t.Fatalf("error before test: %v", err)
 		}
@@ -114,7 +114,7 @@ func TestSearchFileAndFolder(t *testing.T) {
 	})
 
 	t.Run("Search Folder anf File name with space", func(t *testing.T) {
-		app, _, err := testshelper.InitCrowerTests()
+		app, _, err := helper.InitCrowerTests()
 		if err != nil {
 			t.Fatalf("error before test: %v", err)
 		}
